ion: store all datums in encoded forms (#1263)
This changes ion.Datum to represent values in their encoded forms
allowing datums to be passed around as subslices and only converted to
actual values when necessary.

+++ vm/ssa.go
@@ -4935,7 +4935,7 @@ func (c *compilestate) litcmp(v *value, i interface{}) {
 		case float64:
 			d = ion.Float(i)
 		case float32:
-			d = ion.Float(float32(i)) // TODO: maybe don't convert here...
+			d = ion.Float(float64(i)) // TODO: maybe don't convert here...
 		case int64:
 			d = ion.Int(i)
 		case int:
